# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Cuadra API
  version: "1.0.0"
  description: |
    REST API del backend de Cuadra POS.
    Multi-tenant: todos los endpoints filtran por `ownerId` del JWT.
    Auth: Bearer JWT en header `Authorization`.
servers:
  - url: http://localhost:4000
    description: Desarrollo local
  - url: https://cuadra-backend.railway.app
    description: Producción (Railway)

# ─── SECURITY ─────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PromotionCreate:
      type: object
      required: [name, type]
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        type: { type: string, enum: [percentage, fixed, buy_x_get_y] }
        value: { type: number, nullable: true }
        applicable_to:
          type: array
          items: { type: string }
          nullable: true
        min_purchase: { type: number, nullable: true }
        coupon_code: { type: string, nullable: true }
        starts_at: { type: string, format: date-time, nullable: true }
        ends_at: { type: string, format: date-time, nullable: true }
    PromotionResponse:
      allOf:
        - $ref: "#/components/schemas/PromotionCreate"
        - type: object
          properties:
            id: { type: string }
            owner_id: { type: string }
            active: { type: boolean }

    # ── Stock Transfers ──────────────────────────────────────────────────────
...
    SalesSummaryResponse:
      type: object
      properties:
        count: { type: integer }
        revenue: { type: number }
        average: { type: number }
        pending: { type: number }
...
    ProfitReportResponse:
      type: object
      properties:
        revenue: { type: number }
        expenses: { type: number }
        gross_profit: { type: number }
    BcvRateResponse:
      type: object
      properties:
        rate: { type: number, example: 36.5 }
        cached: { type: boolean }
security:
  - bearerAuth: []

# ─── PATHS ────────────────────────────────────────────────────────────────────
paths:
...
  /promotions:
    get:
      summary: Listar promociones activas
      tags: [promotions]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PromotionResponse" }
          description: OK
    post:
      summary: Crear promoción
      tags: [promotions]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PromotionCreate" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PromotionResponse" }
          description: Promoción creada
  /promotions/{promotionId}:
    parameters:
      - name: promotionId
        in: path
        required: true
        schema: { type: string }
    patch:
      summary: Actualizar promoción
      tags: [promotions]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PromotionCreate" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PromotionResponse" }
          description: OK
    delete:
      summary: Desactivar promoción
      tags: [promotions]
      responses:
        "204":
          description: Desactivada

  # ── Stock Transfers ───────────────────────────────────────────────────────────
...
  /reports/sales-summary:
    get:
      summary: Resumen de ventas (revenue, pending, count, avg)
      tags: [reports]
      parameters:
        - name: start_date
          in: query
          schema: { type: string, format: date-time }
        - name: end_date
          in: query
          schema: { type: string, format: date-time }
        - name: location_id
          in: query
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalesSummaryResponse" }
          description: OK
...
  /reports/profit:
    get:
      summary: Reporte de ganancias (revenue - gastos)
      tags: [reports]
      parameters:
        - name: start_date
          in: query
          schema: { type: string, format: date-time }
        - name: end_date
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProfitReportResponse" }
          description: OK

  # ── BCV Rate ──────────────────────────────────────────────────────────────────
  /bcv-rate:
    get:
      summary: Tasa BCV USD/VES (caché 30 min)
      security: []
      tags: [bcv]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BcvRateResponse" }
          description: Tasa actual
        "501":
          description: Scraping no implementado
        "502":
          description: Error fetching BCV